requirements.txt
asgiref==3.6.0
braintree==4.19.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.1
Django==4.1.7
django-appconf==1.0.5
django-ckeditor==6.5.1
django-crispy-forms==2.0
django-image-cropping==1.7
django-jazzmin==2.6.0
django-js-asset==2.0.0
django-mailchimp==0.1.30
django-phonenumber-field==7.0.2
easy-thumbnails==2.8.5
idna==3.4
mailchimp-marketing==3.0.80
phonenumbers==8.13.8
Pillow==9.4.0
pycparser==2.21
PyMySQL==1.0.3
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.15
